refactor(account): improve account module organization and admin operations
